esp_service/requirements.txt
python-dotenv
fastapi
uvicorn[standard]
groq
openai
redis
requests
aiomqtt
pytz